(Build status: Failing, Incomplete commit) Returns an Observable instead of List for API models.
Pros:
- Immutable return type but composable (I believe this is called a
persistent data structure).
- Remove Observable.fromIterable(…) calls everywhere.
- Could open up more doors to the understanding of how to compose the
  software as not thinking about Lists which are time dependent i.e
synchronous but just some asynchronous stream. This means that composing
it with API calls could be more natural inside the model (see
implementation problems).
- No worry about synchronous and asynchronous as pagination and loading
  comments and loading more comments is done by the More.java and
Progress.java and the like so having them in a stream or list makes no
difference.

Cons:
- Can make pulling the APIs out into their own libraries more difficult.
  This is if we want to allow non Rx usage of our libraries. If we force
RxJava usage then this shouldn't be a problem, but I'd have to explain
the usage of this and why I am exposing an Observable instead of List in
the models as it is an odd design decision.

Implementation problems:
Reddit#subredditComments where the tree is being made properly can
instead be put into the model as you can return the mutated Observable
e.g. do the map or flatmap here instead of there. Problem that not
actually defined in model but interface. Could have seperate model (e.g.
ThingsListing with List<Thing<Listing>> inside it). Otherwise could make
Retrofit adapter which converts all Lists to Observables.